Update _entry.twig
Tutorial in Creating Resources "Categories/Topics" section says to `Enter “Post Categories” into the Name field;` and doesn't say to change the handle of the Blog group field to "topics". 

So when `entry.topics.all()` is run the Twig runtime error `Calling unknown method: craft\elements\Entry::topics()` pops up for this line. Unless the tutorial is changed to specify the field handle of "Post Categories" will be "topics" instead of the default "postCategories".